Introjucer: shutdown fix.

+++ modules/juce_core/memory/juce_Memory.h
@@ -97,7 +97,7 @@ inline Type* createCopyIfNotNull (const Type* pointer)     { return pointer != n
  extern JUCE_API void  juceDLL_free (void*);
 
  #define JUCE_LEAK_DETECTOR(OwnerClass)  public:\
-    static void* operator new (size_t sz)           { return juce::juceDLL_malloc ((int) sz); } \
+    static void* operator new (size_t sz)           { return juce::juceDLL_malloc (sz); } \
     static void* operator new (size_t, void* p)     { return p; } \
     static void operator delete (void* p)           { juce::juceDLL_free (p); } \
     static void operator delete (void*, void*)      {}
